pyproject.toml
[tool.poetry]
name = "scribe"
version = "0.1.0"
description = ""
authors = ["yreyricord <yreyricord@argans.co.uk>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.51.0"
pydub = "^0.25.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
